requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
jinja2==3.1.6
numpy==1.26.4
pydantic==2.6.1
python-multipart==0.0.9
sqlalchemy==2.0.27
python-jose==3.3.0
passlib==1.7.4
python-dotenv==1.0.1
psycopg2-binary==2.9.9
bcrypt==4.1.2
httpx==0.26.0
pytest==8.0.0
pytest-asyncio==0.23.5
pandas==2.1.4
openpyxl==3.1.2
scipy==1.11.4
pandas==2.0.3; platform_system!="Windows"
pandas==2.0.3; platform_system=="Windows" and python_version<"3.12"
pandas==2.1.4; platform_system=="Windows" and python_version>="3.12"
scipy==1.11.3; platform_system!="Windows"
scipy==1.11.3; platform_system=="Windows" and python_version<"3.12"
scipy==1.11.4; platform_system=="Windows" and python_version>="3.12" 